requirements.txt
django
channel["daphne"]
